Refactor UI into components and add simulation hook
